feat: rename progressbar to spinner
it is neither a bar neither indicates progress.
