+++ cmd/embedded-cluster/install.go
@@ -21,9 +21,9 @@ import (
 	"github.com/replicatedhq/embedded-cluster/pkg/helpers"
 	"github.com/replicatedhq/embedded-cluster/pkg/metrics"
 	"github.com/replicatedhq/embedded-cluster/pkg/preflights"
-	"github.com/replicatedhq/embedded-cluster/pkg/progressbar"
 	"github.com/replicatedhq/embedded-cluster/pkg/prompts"
 	"github.com/replicatedhq/embedded-cluster/pkg/release"
+	"github.com/replicatedhq/embedded-cluster/pkg/spinner"
 )
 
 // ErrNothingElseToAdd is an error returned when there is nothing else to add to the
@@ -77,7 +77,7 @@ func runHostPreflights(c *cli.Context) error {
 	if len(hpf.Collectors) == 0 && len(hpf.Analyzers) == 0 {
 		return nil
 	}
-	pb := progressbar.Start()
+	pb := spinner.Start()
 	pb.Infof("Running host preflights on node")
 	output, err := preflights.Run(c.Context, hpf)
 	if err != nil {
@@ -211,9 +211,9 @@ func installK0s(c *cli.Context) error {
 // waitForK0s waits for the k0s API to be available. We wait for the k0s socket to
 // appear in the system and until the k0s status command to finish.
 func waitForK0s(ctx context.Context) error {
-	pb := progressbar.Start()
-	defer pb.Close()
-	pb.Infof("Waiting for %s node to be ready", defaults.BinaryName())
+	loading := spinner.Start()
+	defer loading.Close()
+	loading.Infof("Waiting for %s node to be ready", defaults.BinaryName())
 	var success bool
 	for i := 0; i < 30; i++ {
 		time.Sleep(2 * time.Second)
@@ -230,7 +230,7 @@ func waitForK0s(ctx context.Context) error {
 	if _, err := runCommand(defaults.K0sBinaryPath(), "status"); err != nil {
 		return fmt.Errorf("unable to get status: %w", err)
 	}
-	pb.Infof("Node installation finished")
+	loading.Infof("Node installation finished")
 	return nil
 }
